requirements.txt
discord.py[voice]
discord.py
discord-py-slash-command
Pillow
requests
python-dotenv
discord-interactions
asyncio
